Extract port detection to separate service.
