+++ src/services/tyeProcessProvider.ts
@@ -3,8 +3,8 @@
 
 import { Observable, timer } from 'rxjs'
 import { distinctUntilChanged, first, switchMap } from 'rxjs/operators';
+import { PortProvider } from './portProvider';
 import { ProcessProvider } from './processProvider';
-import * as netstat from 'node-netstat';
 
 export interface TyeProcess {
     pid: number;
@@ -14,59 +14,6 @@ export interface TyeProcess {
 export interface TyeProcessProvider {
     readonly processes: Observable<TyeProcess[]>;
     getProcesses(): Promise<TyeProcess[]>;
-}
-
-netstat.commands.darwin = {
-    cmd: 'lsof',
-    args: ['-Pn', '-i4', '-sTCP:LISTEN']
-};
-
-// NOTE: The TS definition incorrectly asserts that `darwin` is const.
-// TODO: Update the TS definitions.
-(netstat.parsers.darwin as unknown) = (line: string, callback: (item: netstat.ParsedItem) => void) => {
-    const parts = line.split(/\s/).filter(String);
-    if (!parts.length || (parts.length != 9 && parts.length != 10 )) {
-        return;
-    }
-
-    let state = parts[9] || '';
-
-    if (state.length >= 2 && state[0] === '(' && state[state.length - 1] === ')') {
-        state = state.slice(1, state.length - 1);
-    }
-
-    const item = {
-        protocol: parts[7],
-        local: parts[8],
-        remote: '',
-        state,
-        pid: parts[1]
-    };
-
-    return callback(netstat.utils.normalizeValues(item));
-};
-
-function netstatAsync(options: Omit<netstat.Options, 'done'>): Promise<netstat.ParsedItem[]> {
-    return new Promise(
-        (resolve, reject) => {
-            const items: netstat.ParsedItem[] = [];
-
-            netstat(
-                {
-                    ...options,
-                    done:
-                        error => {
-                            if (error) {
-                                reject(new Error(error));
-                            } else {
-                                resolve(items);
-                            }
-                        }
-                },
-                item => {
-                    items.push(item);
-                });
-        });
 }
 
 function tyeProcessComparer(x: TyeProcess, y: TyeProcess): boolean {
@@ -98,7 +45,9 @@ type ProspectiveTyeProcess = WithOptional<TyeProcess, 'dashboardPort'>;
 export default class LocalTyeProcessProvider implements TyeProcessProvider {
     private readonly _processes: Observable<TyeProcess[]>;
 
-    constructor(private readonly processProvider: ProcessProvider) {
+    constructor(
+        private readonly portProvider: PortProvider,
+        private readonly processProvider: ProcessProvider) {
         this._processes =
             // TODO: Make interval configurable.
             timer(0, 2000)
@@ -139,9 +88,9 @@ export default class LocalTyeProcessProvider implements TyeProcessProvider {
     }
 
     private async getPortForProcess(pid: number): Promise<ProspectiveTyeProcess> {
-        const items = await netstatAsync({ filter: { pid, protocol: 'tcp' } });
+        const ports = await this.portProvider.getPortsForProcess(pid);
 
-        return { pid, dashboardPort: items[0]?.local?.port ?? undefined }
+        return { pid, dashboardPort: ports[0] }
     }
 }
